test(bigquery-jdbc): relocate testCalendarConversions from BigQueryParameterHandlerTest to FieldValueTypeBigQueryCoercionUtilityTest
